requirements.txt
google-auth-oauthlib>=0.4.6
google-api-python-client>=2.86.0
google-auth-httplib2>=0.1.0
pytz>=2023.3
python-dateutil>=2.8.2
pandas>=2.0.0
click>=8.1.3 